Code review commentaar verwerkt
