+++ backend/sparql/src/test/java/org/dotwebstack/framework/backend/sparql/SparqlBackendSourceTest.java
@@ -3,12 +3,19 @@ package org.dotwebstack.framework.backend.sparql;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertThat;
 
-import org.dotwebstack.framework.test.DBEERPEDIA;
+import org.dotwebstack.framework.backend.Backend;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
 
+@RunWith(MockitoJUnitRunner.class)
 public class SparqlBackendSourceTest {
+
+  @Mock
+  Backend backend;
 
   @Rule
   public ExpectedException thrown = ExpectedException.none();
@@ -17,10 +24,10 @@ public class SparqlBackendSourceTest {
   public void builder() {
     // Act
     SparqlBackendSource backendSource =
-        new SparqlBackendSource.Builder(DBEERPEDIA.BACKEND, "myQuery").build();
+        new SparqlBackendSource.Builder(backend, "myQuery").build();
 
     // Assert
-    assertThat(backendSource.getBackendReference(), equalTo(DBEERPEDIA.BACKEND));
+    assertThat(backendSource.getBackend(), equalTo(backend));
     assertThat(backendSource.getQuery(), equalTo("myQuery"));
   }
