+++ core/src/main/java/org/dotwebstack/framework/backend/BackendSource.java
@@ -1,7 +1,5 @@
 package org.dotwebstack.framework.backend;
 
-import org.eclipse.rdf4j.model.IRI;
-
 public interface BackendSource {
-  IRI getBackendReference();
+  Backend getBackend();
 }
