+++ core/src/test/java/org/dotwebstack/framework/informationproduct/InformationProductLoaderTest.java
@@ -10,7 +10,6 @@ import static org.mockito.Mockito.when;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import org.dotwebstack.framework.InformationProduct;
 import org.dotwebstack.framework.Registry;
 import org.dotwebstack.framework.backend.Backend;
 import org.dotwebstack.framework.backend.BackendSource;
